+++ NEWS.md
@@ -1,9 +1,10 @@
 
 # DIDdesign 0.2.3
 
-+ Update `summary()` function. Now it has an option to print all the results.  
++ Updated `summary()` function. Now it has an option to print all the results.
 + Fixed bugs on how formula input is handled.
 + Added `malesky2014` dataset.
++ Updated `did_plot()` function. Now it plots confidence intervals along with mean outcome when `ci = TRUE`.
 
 # DIDdesign 0.2.2